# api/swagger.yaml
openapi: 3.0.0
servers:
  - description: BOG Production
    url: https://bog.simplenight.com/api/v1
  - description: Simplenight Live Sandbox
    url: https://simplenight-api-278418.ue.r.appspot.com/api/v1
  - description: Localhost
    url: http://localhost:8000/api/v1
info:
  version: "2.0"
  title: Simplenight API
  "x-logo": {
    "url": "simplenight-logo.png",
    "backgroundColor": "#FFFFFF",
    "altText": "Simplenight logo"
  }
  description: |
    This document describes the Simplenight API.

    * The Hotels API supports availability and booking of hotels alone. This API is now documented, in favor of the
      multi-product API.

    * The Locations API provides a list of Simplenight-supported locations, and convenience APIs for auto-complete.

    * The Multi-Product API supports Hotel, Activity and Car searches.

    Requests can be tried live in our [Swagger UI interface](/index.html)
...
security:
  - ApiKeyAuth: [ ]
paths:
  /hotels/google/search-by-id:
    post:
      tags:
        - BookOnGoogle
      summary: Book on Google Hotel Availability.
      description: This endpoint performs a single hotel search, with an API response suitable for Book on Google
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HotelSearchRequest'
      responses:
        '200':
          description: List of matching locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HotelSearchResponse'
        '400':
          description: bad input parameter
        '404':
          description: No locations found for input
  /hotels/booking:
    post:
      tags:
        - Hotel
      summary: Submit a booking request and receive a Reservation in response.
      description: Collect a customer record, payment information and itinerary for booking and return a Reservation on success.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HotelBookingRequest'
      responses:
        '200':
          description: Reservation on success, or else an Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotelBookingResponse'
        '400':
          description: Bad request
        '500':
          description: Internal error while completing a booking
  /hotels/cancel:
    post:
      tags:
        - Hotel
      summary: Lookup cancel info
      description: Lookup a booking with a Booking ID and Lead Traveler Last Name, and return current Cancellation Policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelRequest'
      responses:
        '200':
          description: The current matching cancellation policy, or an error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelResponse'
        '400':
          description: Bad request
        '500':
          description: Internal error while completing a booking
  /hotels/cancel-confirm:
    post:
      tags:
        - Hotel
      summary: Lookup cancel info
      description: Lookup a booking with a Booking ID and Lead Traveler Last Name, and return current Cancellation Policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelRequest'
      responses:
        '200':
          description: The current matching cancellation policy, or an error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmCancelResponse'
        '400':
          description: Bad request
        '500':
          description: Internal error while completing a booking
  /hotels/search-by-id:
    post:
      tags:
        - Hotel
      summary: Search for a specific hotel
      description: Search for a specific hotel, with a Hotel ID, travel dates, and an occupancy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HotelSearchRequest'
      responses:
        '200':
          description: Matching hotel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotelSearchResponse'
        '400':
          description: bad input parameter
        '404':
          description: No locations found for input
  /hotels/search-by-location:
    post:
      tags:
        - Hotel
      summary: Search for hotels in a location
      description: Using a Simplenight city ID, multi for hotels
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HotelLocationSearchRequest'
      responses:
        '200':
          description: List of matching Hotels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HotelSearchResponse'
        '400':
          description: bad input parameter
        '404':
          description: No locations found for input
  /multi/search:
    post:
      tags:
        - MultiProduct
      summary: Search
      operationId: multiProductSearch
      description: |
        Availability search for multiple products.  Currently supports Hotel, Activity, Transportation, and Dining.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MultiProductSearchRequest'
      responses:
        '200':
          description: List of matching products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MultiProductSearchResponse'
        '400':
          description: bad input parameter
        '404':
          description: No locations found for input
  /multi/details:
    post:
      tags:
        - MultiProduct
      summary: Activity Details
      description: |
        Returns the specific details an activity product is available.  Using a `code` from a [Search](#operation/multiProductSearch),
        return the specific dates an activity is available between a `date_from` and a `date_to`.
      operationId: activityDetails
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityDetailsRequest'
      responses:
        '200':
          description: List of matching products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActivityDetailsResponse'
        '400':
          description: bad input parameter
        '404':
          description: No locations found for input
  /multi/variants:
    post:
      tags:
        - MultiProduct
      summary: Activity Tickets
      operationId: activityTickets
      description: |
        Returns the specific times an activity is available, along with prices for specific ticket types (e.g., adults,=
        children, senior, etc).

        This API returns an associative map, keyed by the activity times.  For example,

            {
              "variants": {
                "12:00": [
                  {
                    "code": "2636",
                    "name": "Adult",
                    "description": "Age: 25 - 64",
                    "price": 16.78,
                    "capacity": 20,
                    "additional": {}
                  }
                ],
                "13:00": [
                  {
                    "code": "2636",
                    "name": "Adult",
                    "description": "Age: 25 - 64",
                    "price": 16.78,
                    "capacity": 20,
                    "additional": {}
                  }
                ]
              }
            }

          When the key is `whole_day`, the activity has a single start time, and all products apply to that start time.

            {
              "variants": {
                "whole_day": [
                  {
                    "code": "2636",
                    "name": "Adult",
                    "description": "Age: 25 - 64",
                    "price": 16.78,
                    "capacity": 20,
                    "additional": {}
                  }
                ]
              }
            }

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityTicketRequest'
      responses:
        '200':
          description: List of activity ticket options
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActivityTicketResponse'
        '400':
          description: bad input parameter
        '404':
          description: No locations found for input
  /multi/book:
    post:
      tags:
        - MultiProduct
      summary: Multi-Product Book
      operationId: multiProductBook
      description: |
        Book multiple products in a single transaction.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MultiProductBookingRequest'
      responses:
        '200':
          description: List of matching products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MultiProductBookingResponse'
        '400':
          description: bad input parameter
        '404':
          description: No locations found for input
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: |
        Simplenight uses a API Key passed as a header value for authentication.  Simplenight must generate this API Key
        for you.  API Keys identify your organization, and enable separate configurations and feature flags.
  schemas:
...
    HotelSearchResponse:
      type: object
      required:
        - hotel_code
        - start_date
        - end_date
        - occupancy
        - room_types
        - rate_plans
        - room_rates
        - hotel_details
        - average_nightly_rate
        - average_nightly_tax
        - average_nightly_base
      properties:
        hotel_code:
          type: string
          example: 12345
          description: A SimpleNights Hotel ID
        start_date:
          type: string
          format: date
          description: Checkin date
        end_date:
          type: string
          format: date
          description: Checkout date
        occupancy:
          $ref: '#/components/schemas/Occupancy'
        room_types:
          type: array
          items:
            $ref: '#/components/schemas/RoomType'
        rate_plans:
          type: array
          items:
            $ref: '#/components/schemas/RatePlan'
        room_rates:
          type: array
          items:
            $ref: '#/components/schemas/RoomRate'
        hotel_details:
          $ref: '#/components/schemas/HotelDetails'
        average_nightly_rate:
          type: number
          description: The minimum total nightly rate.  Equal to minimum rate total / room_nights
          example: 107.29
        average_nightly_tax:
          type: number
          description: The minimum total nightly taxes.  Equal to the minimum tax total / room_nights
        average_nightly_base:
          type: number
          description: The minimum total nightly taxes.  Equal to the minimum base total / room_nights
        error:
          $ref: '#/components/schemas/AvailabilityError'
    HotelSearchRequest:
      type: object
      required:
        - hotel_id
        - start_date
        - end_date
        - occupancy
      properties:
        hotel_id:
          type: string
          example: SJGH25
          description: Simplenight Hotel ID
        start_date:
          type: string
          format: date
          description: The checkin date
        end_date:
          type: string
          format: date
          description: The checkout date
        occupancy:
          $ref: '#/components/schemas/Occupancy'
        language:
          type: string
          minLength: 2
          maxLength: 2
          example: en
          description: Optional. ISO-formatted 2-character language code for hotel details
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: USD
          description: Optional.  ISO-format 3-character currency code
    HotelLocationSearchRequest:
      type: object
      required:
        - location_id
        - start_date
        - end_date
        - occupancy
      properties:
        location_id:
          type: string
          example: "19823"
          description: Simplenight Location ID
        start_date:
          type: string
          format: date
          description: The checkin date
        end_date:
          type: string
          format: date
          description: The checkout date
        occupancy:
          $ref: '#/components/schemas/Occupancy'
        language:
          type: string
          minLength: 2
          maxLength: 2
          example: en
          description: Optional. ISO-formatted 2-character language code for hotel details
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: USD
          description: Optional.  ISO-format 3-character currency code
    Occupancy:
      type: object
      description: The number of adults and children
      required:
        - adults
        - children
      properties:
        adults:
          type: integer
          example: 2
          description: The number of adults per room
        children:
          type: integer
          example: 0
          description: The number of children per room
        num_rooms:
          type: integer
          example: 1
          description: The number of rooms requested.  Applies to the same room type.
...
    RoomType:
      type: object
      description: The types of physical rooms available
      required:
        - code
        - name
      properties:
        code:
          type: string
          description: The unique ID for the room type
          example: H214A
        name:
          type: string
          description: The descriptive name for the room type
          example: Run of the House
        description:
          type: string
          description: The long description of the room type
          example: The best available room at the time of checkin
        basic_amenities:
          type: array
          items:
            $ref: '#/components/schemas/Amenity'
        photos:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
        capacity:
          $ref: '#/components/schemas/Capacity'
        bed_types:
          $ref: '#/components/schemas/BedTypes'
        unstructured_policies:
          type: string
          description: Textual description of hotel rate policies
...
    RatePlan:
      type: object
      description: Rate plans corresponding to specific hotel bundles and services
      required:
        - code
        - name
        - guarantee_type
        - cancellation_policy
      properties:
        code:
          type: string
          example: S245GK
          description: The unique ID for the room rate plan
        name:
          type: string
          example: 5th Night Free
          description: Textual description of the rate plan
        description:
          type: string
          example: Stay 4 nights and get the 4th night absolutely free
          description: Long description of the rate plan
        basic_amenities:
          $ref: '#/components/schemas/Amenity'
        cancellation_policy:
          $ref: '#/components/schemas/CancellationPolicy'
        unstructured_policies:
          type: string
          example: Cancellations after 24 hours incur a 100% charge
          description: Textual description of hotel policies
...
    AvailabilityError:
      required:
        - type
        - message
      properties:
        type:
          $ref: '#/components/schemas/AvailabilityErrorType'
        message:
          type: string
          example: Error while searching hotel supplier foo
          description: Unstructed, detailed error message
...
    HotelBookingRequest:
      required:
        - api_version
        - transaction_id
        - hotel_id
        - room_code
        - language
        - customer
        - traveler
        - payment
      properties:
        api_version:
          type: integer
          example: 1
          description: The API version of this booking request message
        transaction_id:
          type: string
          example:
            3ef11392-a96a-423f-8bc6-4207d957a545
          description: Unique ID for this booking transaction
        hotel_id:
          type: string
          example: AHY14Z
          description: The SimpleNight Hotel ID of the property being booked
        room_code:
          type: string
          example: X2HD4J
          description: The room code associated with the room rate to book
        language:
          type: string
          example: en_US
          description: The language of the request, for hotel details and property descriptions
        customer:
          $ref: '#/components/schemas/Customer'
        traveler:
          $ref: '#/components/schemas/Traveler'
        payment:
          $ref: '#/components/schemas/Payment'
        additional_info:
          type: string
          example:
            Twin bed please
          description: Additional information for hotel booking request
...
    HotelBookingResponse:
      type: object
      description: The booking response for a successful booking
      required:
        - api_version
        - transaction_id
        - status
        - reservation
      properties:
        api_version:
          type: integer
          example: 2
          description: The API version of the booking response
        transaction_id:
          type: string
          example: 3ef11392-a96a-423f-8bc6-4207d957a545
          description: Echoed transaction_id response
        status:
          $ref: '#/components/schemas/Status'
        reservation:
          $ref: '#/components/schemas/Reservation'
    Status:
      description: Success/Failure indicator for a booking
      type: object
      enum:
        - SUCCESS
        - FAILURE
...
    CancelRequest:
      type: object
      description: Cancellation information request.  Returns the current matching cancellation policy
      required:
        - booking_id
        - last_name
      properties:
        booking_id:
          type: string
          example: eb276848-9093-4f50-9185-1515d658586c
          description: The booking ID returned in the booking message
        last_name:
          type: string
          example: Doe
          description: The lead traveler last name
...
    CancelResponse:
      type: object
      description: The current matching cancellation policy
      properties:
        is_cancellable:
          type: boolean
          description: Whether, or not the booking is currently cancellable
        booking_id:
          type: string
          description: Booking ID from the Booking Response
          format: uuid
        itinerary:
          $ref: '#/components/schemas/Itinerary'
        details:
          $ref: '#/components/schemas/CancelDetails'
    ConfirmCancelResponse:
      type: object
      description: Confirmation that a booking was cancelled
      properties:
        booking_id:
          type: string
          description: The booking ID from the booking response
        last_name:
          type: string
          description: The lead traveler last name, used for verification purposes (exact match)
...
    ActivityDetailsRequest:
      type: object
      description: Parameters for activity details, to return the specific dates an activity is available.
      properties:
        code:
          type: string
          description: The same code returned in the activity search response.
          example: 1a87b6d7
        date_from:
          type: string
          format: date
          description: The start date to search for, inclusive
        date_to:
          type: string
          format: date
          description: The end date to search for, inclusive
...
    ActivityDetailsResponse:
      type: object
      description: |
        Contains the specific dates an activity is available, along with details like its cancellation policy
      properties:
        code:
          type: string
          description: The identifier for this activity product
          example: 1a87b6d7
        type:
          type: string
          description: An enum value representing the type of activity
        categories:
          type: array
          items:
            type: string
        timezone:
          type: string
          description: The local time zone for the activity
          example: America/Los_Angeles
        images:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
        contact:
          $ref: '#/components/schemas/BusinessContact'
        location:
          $ref: '#/components/schemas/ActivityLocation'
        availabilities:
          type: array
          description: The specific days a property is available.  Times are available in the Variants response.
          items:
            type: string
            format: date
        policies:
          type: array
          description: Free-form policies for the activity
          items:
            type: string
        cancellations:
          $ref: '#/components/schemas/CancellationPolicy'
    MultiProductSearchRequest:
      type: object
      description: |
        Parameters to search for multiple products.  Within this request, there can be multiple sub-search types,
        such as an ActivitySearch, HotelSearch, or TransportationSearch.  The results for each product type will be returned
        in a single response.
      properties:
        product_types:
          type: array
          enum:
            - activities
            - hotels
            - transporation
            - dining
        activity_search:
          $ref: '#/components/schemas/ActivitySearchRequest'
        hotel_search:
          $ref: '#/components/schemas/HotelSearchRequest'
    MultiProductSearchResponse:
      type: object
      description: |
        Parameters to search for multiple products.  Within this request, there can be multiple sub-search types,
        such as an ActivitySearch, HotelSearch, or TransportationSearch.  The results for each product type will be returned
        in a single response.
      properties:
        product_types:
          type: array
          enum:
            - activities
            - hotels
            - transportation
            - dining
        activity_search:
          $ref: '#/components/schemas/HotelSearchResponse'
        hotel_search:
          $ref: '#/components/schemas/ActivitySearchResponse'
    ActivityTicketRequest:
      type: object
      description: |
        The `code` from an [Availability](#operation/multiProductAvailability) call, and an `activity_date` from an
        [Activity Details](#operation/activityDetails) call.
      properties:
        code:
          type: string
          example: 0e29b9ef
        activity_date:
          type: string
          format: date
    ActivityTicketResponse:
      type: object
      description: |
        The available ticket types for a specific activity on a specific date.

      properties:
        variants:
          additionalProperties:
            type: object
            example: foo
            properties:
              code:
                type: string
              name:
                type: string
                example: Adult
              description:
                type: string
                example: "Age: 25 - 64"
              price:
                type: number
                example: 16.78
              capacity:
                type: integer
                example: 20
              additional:
                description: Additional policy information
                additionalProperties:
                  type: string
...
    MultiProductBookingRequest:
      type: object
      description: Booking request supporting multiple products in a single transaction.
      properties:
        api_version:
          type: integer
          example: 1
          description: The API version of the request
        transaction_id:
          type: string
          format: uuid
          description: Arbitrary string used to identify a request, and prevent duplicate bookings
        language:
          type: string
          description: a locale string for the customer
          example: en_US
        customer:
          $ref: '#/components/schemas/Customer'
        traveler:
          $ref: '#/components/schemas/Traveler'
        payment:
          $ref: '#/components/schemas/Payment'
        hotel_booking:
          $ref: '#/components/schemas/MultiProductHotelBookingRequest'
        activity_booking:
          $ref: '#/components/schemas/MultiProductActivityBookingRequest'
        additional_info:
          type: string
          description: Additional information for booking request
          example: Twin bed please
    MultiProductBookingResponse:
      type: object
      description: Booking response for multiple product types, booked in a single transaction/request.
      properties:
        api_version:
          type: integer
          example: 1
          description: The API version of the response
        transaction_id:
          type: string
          format: uuid
          description: The echoed transaction ID from the booking request
        status:
          $ref: '#/components/schemas/Status'
        booking_id:
          type: string
          description: The Simplenight Record Locator for the completed booking
          example: R75QY0B9
        hotel_reservation:
          $ref: '#/components/schemas/HotelBookingResponse'
        activity_booking:
          $ref: '#/components/schemas/MultiProductActivityBookingResponse'